Consolidate concurrent and evaluator component into `RuntimeModule`
