+++ src/main/scala/com/hhandoko/aktiform/app/runner/AktiformApplication.scala
@@ -7,12 +7,7 @@ import org.springframework.context.annotation.Import
 
 import com.hhandoko.aktiform.app.config.ResourcesConfig
 import com.hhandoko.aktiform.app.controller.{FormController, HelloWorldController, HomeController}
-import com.hhandoko.aktiform.app.module.{
-  EvaluatorModule,
-  ExecutionContextModule,
-  PolyglotContextModule,
-  StandaloneProviderModule
-}
+import com.hhandoko.aktiform.app.module.{RuntimeModule, PolyglotContextModule, StandaloneProviderModule}
 
 /** Main application class (Spring Application runner).
   *
@@ -28,9 +23,8 @@ import com.hhandoko.aktiform.app.module.{
 @Import(
   value = Array(
     // Modules
-    classOf[EvaluatorModule],
-    classOf[ExecutionContextModule],
     classOf[PolyglotContextModule],
+    classOf[RuntimeModule],
     classOf[StandaloneProviderModule],
     // Controllers
     classOf[FormController],
